renames dir
